requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
distlib==0.3.3
Django==3.2.4
filelock==3.4.0
mysqlclient==2.1.0
pipenv==2018.10.13
platformdirs==2.3.0
python-telegram-bot==13.6
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzdata==2021.5
tzlocal==2.1
virtualenv==20.8.1
virtualenv-clone==0.5.7
